Added Recipe class.

+++ DrinkDispenser/DrinkDispenser.cs
@@ -2,7 +2,7 @@
 
 public class DrinkDispenser
 {
-    public static double GetDrinkPrice(string recipeName)
+    public static decimal GetDrinkPrice(string recipeName)
     {
         return 0;
     }
